up libcoco
Former-commit-id: 3cfd6b3309b5e4d0e36fe9dc7db8dc5730516ae5

+++ epbin/src/epe/PackAnimation.h
@@ -16,7 +16,7 @@ namespace epe
 class PackAnimation : public IPackNode
 {
 public:
-	PackAnimation(lua_State* L, uint16_t id);
+	PackAnimation(lua_State* L, uint32_t id);
 	virtual ~PackAnimation();
 
 	virtual size_t Size() const;
