+++ epbin/src/epe/PackPannel.h
@@ -15,7 +15,7 @@ namespace epe
 class PackPannel : public IPackNode
 {
 public:
-	PackPannel(lua_State* L, uint16_t id);
+	PackPannel(lua_State* L, uint32_t id);
 
 	virtual size_t Size() const;
 	virtual void Store(std::ofstream& fout) const;
