+++ epbin/src/epe/PackPolygon.h
@@ -15,7 +15,7 @@ namespace epe
 class PackPolygon : public IPackNode
 {
 public:
-	PackPolygon(lua_State* L, uint16_t id);
+	PackPolygon(lua_State* L, uint32_t id);
 
 	virtual size_t Size() const;
 	virtual void Store(std::ofstream& fout) const;
